+++ src/main/java/com/alexey/dao/FilmmakerDao.java
@@ -8,9 +8,9 @@ import java.util.List;
 
 public interface FilmmakerDao {
     List<Filmmaker> getAll() throws DataAccessException;
-    Filmmaker getById(Long id) throws DataAccessException;
-    Long insertFilmmaker(Filmmaker filmmaker) throws DataAccessException;
+    Filmmaker getById(Integer id) throws DataAccessException;
+    Integer insertFilmmaker(Filmmaker filmmaker) throws DataAccessException;
     void updateFilmmaker(Filmmaker filmmaker) throws DataAccessException;
-    void deleteFilmmaker(Long id) throws DataAccessException;
-   // List<Movie> getHisMovies(Long id) throws DataAccessException;
+    void deleteFilmmaker(Integer id) throws DataAccessException;
+   // List<Movie> getHisMovies(Integer id) throws DataAccessException;
 }
