+++ src/main/java/com/alexey/dao/MovieDao.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface MovieDao {
     List<Movie> getAll() throws DataAccessException;
-    Movie getById(Long id) throws DataAccessException;
-    List<Movie> getMoviesByFilmmaker(Long id) throws DataAccessException;
-    Long insertMovie(Movie movie) throws DataAccessException;
+    Movie getById(Integer id) throws DataAccessException;
+    List<Movie> getMoviesByFilmmaker(Integer id) throws DataAccessException;
+    Integer insertMovie(Movie movie) throws DataAccessException;
     void updateMovie(Movie movie) throws DataAccessException;
-    void deleteMovie(Long id) throws DataAccessException;
+    void deleteMovie(Integer id) throws DataAccessException;
 }
